move testdata production if variable has __on_consumer__ = True
required alot of refactoring to avoid circular
imports. which also resulted in a "cleaner"
import path for many other things.
