+++ .github/workflows/code-quality.yaml
@@ -1,7 +1,6 @@
 name: code quality
 
 on:
-  push:
   pull_request:
     types: [opened]
   workflow_dispatch:
