+++ grizzly/testdata/utils.py
@@ -1,10 +1,21 @@
-from typing import Optional, List, Dict, Any, Tuple, Set
+import logging
+
+from typing import Optional, List, Dict, Any, Tuple, Set, cast
+from collections import namedtuple
+from os import environ
+
+from jinja2 import Template, Environment
+from jinja2.meta import find_undeclared_variables
+from locust.exception import StopUser
 
 from ..context import GrizzlyContext
 from ..task import RequestTask
-from ..types import TestdataType
-from .variables import load_variable
+from ..types import TestdataType, GrizzlyDictValueType, GrizzlyDict
+from ..utils import merge_dicts
 from .ast import get_template_variables
+
+
+logger = logging.getLogger(__name__)
 
 
 def initialize_testdata(sources: Optional[List[RequestTask]]) -> Tuple[TestdataType, Set[str]]:
@@ -39,10 +50,114 @@ def _get_variable_value(name: str) -> Tuple[Any, Set[str]]:
 
     if '.' in name:
         [variable_type, variable_name] = name.split('.', 1)
-        variable = load_variable(variable_type)
+        variable = GrizzlyDict.load_variable(variable_type)
         external_dependencies = variable.__dependencies__
-        value = variable(variable_name, default_value)
+        if getattr(variable, '__on_consumer__', False):
+            value = cast(Any, '__on_consumer__')
+        else:
+            value = variable(variable_name, default_value)
     else:
         value = default_value
 
     return value, external_dependencies
+
+
+def transform(data: Dict[str, Any], objectify: Optional[bool] = True) -> Dict[str, Any]:
+    testdata: Dict[str, Any] = {}
+
+    for key, value in data.items():
+        if '.' in key:
+            if value == '__on_consumer__':
+                grizzly = GrizzlyContext()
+                [variable_type, variable_name] = key.split('.', 1)
+                variable_class_type = GrizzlyDict.load_variable(variable_type)
+                try:
+                    initial_value = grizzly.state.variables.get(key, None)
+                    variable_instance = variable_class_type(variable_name, initial_value)
+                except ValueError as e:
+                    if 'object already has attribute' in str(e):
+                        variable_instance = variable_class_type.get()
+                    else:
+                        raise e
+
+                try:
+                    value = variable_instance[variable_name]
+                except Exception as e:
+                    logger.error(f'{key}: {str(e)}', exc_info=grizzly.state.verbose)
+                    raise StopUser()
+
+            paths: List[str] = key.split('.')
+            variable = paths.pop(0)
+            path = paths.pop()
+            struct = {path: value}
+            paths.reverse()
+
+            for path in paths:
+                struct = {path: {**struct}}
+
+            if variable in testdata:
+                testdata[variable] = merge_dicts(testdata[variable], struct)
+            else:
+                testdata[variable] = {**struct}
+        else:
+            testdata[key] = value
+
+    if objectify:
+        return _objectify(testdata)
+    else:
+        return testdata
+
+
+def _objectify(testdata: Dict[str, Any]) -> Dict[str, Any]:
+    for variable, attributes in testdata.items():
+        if not isinstance(attributes, dict):
+            continue
+
+        attributes = _objectify(attributes)
+        testdata[variable] = namedtuple('Testdata', attributes.keys())(**attributes)
+
+    return testdata
+
+
+def create_context_variable(grizzly: GrizzlyContext, variable: str, value: str) -> Dict[str, Any]:
+    casted_value = resolve_variable(grizzly, value)
+
+    variable = variable.lower().replace(' ', '_').replace('/', '.')
+
+    return transform({variable: casted_value}, objectify=False)
+
+
+def resolve_variable(grizzly: GrizzlyContext, value: str, guess_datatype: Optional[bool] = True) -> GrizzlyDictValueType:
+    if len(value) < 1:
+        return value
+
+    resolved_variable: GrizzlyDictValueType
+    if '{{' in value and '}}' in value:
+        template = Template(value)
+        j2env = Environment(autoescape=False)
+        template_parsed = j2env.parse(value)
+        template_variables = find_undeclared_variables(template_parsed)
+
+        for template_variable in template_variables:
+            assert template_variable in grizzly.state.variables, f'value contained variable "{template_variable}" which has not been set'
+
+        resolved_variable = template.render(**grizzly.state.variables)
+    elif len(value) > 4 and value[0] == '$':
+        if value[0:5] == '$conf':
+            variable = value[7:]
+            assert variable in grizzly.state.configuration, f'configuration variable "{variable}" is not set'
+            resolved_variable = grizzly.state.configuration[variable]
+        elif value[0:4] == '$env':
+            variable = value[6:]
+            env_value = environ.get(variable, None)
+            assert env_value is not None, f'environment variable "{variable}" is not set'
+            resolved_variable = env_value
+        else:
+            raise ValueError(f'{value.split("::", 1)[0]} is not implemented')
+    else:
+        resolved_variable = value
+
+    if guess_datatype:
+        resolved_variable = GrizzlyDict.guess_datatype(resolved_variable)
+
+    return resolved_variable
