--- grizzly/testdata/variables/integer.py
@@ -1,6 +1,0 @@
-from . import AtomicVariable
-
-
-class AtomicInteger(AtomicVariable[int]):
-    __base_type__ = int
-    pass
