+++ tests/test_grizzly/test_utils.py
@@ -1,6 +1,4 @@
-import os
-
-from typing import Any, Callable, Optional, Type, Dict, cast
+from typing import Optional, Type, cast
 from types import FunctionType
 
 import pytest
@@ -18,10 +16,6 @@ from grizzly.utils import (
     create_task_class_type,
     create_user_class_type,
     fail_direct,
-    create_context_variable,
-    resolve_variable,
-    _objectify,
-    transform,
     in_correct_section
 )
 from grizzly.types import RequestMethod
@@ -353,292 +347,6 @@ def test_create_task_class_type() -> None:
         scenario = GrizzlyContextScenario()
         scenario.name = 'A scenario description'
         create_task_class_type('DoesNotExistInGrizzlyScenariosModel', scenario)
-
-
-def test_create_context_variable() -> None:
-    grizzly = GrizzlyContext()
-
-    try:
-        assert create_context_variable(grizzly, 'test.value', '1') == {
-            'test': {
-                'value': 1,
-            }
-        }
-
-        assert create_context_variable(grizzly, 'test.value', 'trUe') == {
-            'test': {
-                'value': True,
-            }
-        }
-
-        assert create_context_variable(grizzly, 'test.value', 'AZURE') == {
-            'test': {
-                'value': 'AZURE',
-            }
-        }
-
-        assert create_context_variable(grizzly, 'test.value', 'HOST') == {
-            'test': {
-                'value': 'HOST'
-            }
-        }
-
-        with pytest.raises(AssertionError):
-            create_context_variable(grizzly, 'test.value', '$env::HELLO_WORLD')
-
-        os.environ['HELLO_WORLD'] = 'environment variable value'
-        assert create_context_variable(grizzly, 'test.value', '$env::HELLO_WORLD') == {
-            'test': {
-                'value': 'environment variable value',
-            }
-        }
-
-        os.environ['HELLO_WORLD'] = 'true'
-        assert create_context_variable(grizzly, 'test.value', '$env::HELLO_WORLD') == {
-            'test': {
-                'value': True,
-            }
-        }
-
-        os.environ['HELLO_WORLD'] = '1337'
-        assert create_context_variable(grizzly, 'test.value', '$env::HELLO_WORLD') == {
-            'test': {
-                'value': 1337,
-            }
-        }
-
-        with pytest.raises(AssertionError):
-            create_context_variable(grizzly, 'test.value', '$conf::test.auth.user.username')
-
-        grizzly.state.configuration['test.auth.user.username'] = 'username'
-        assert create_context_variable(grizzly, 'test.value', '$conf::test.auth.user.username') == {
-            'test': {
-                'value': 'username',
-            }
-        }
-
-        grizzly.state.configuration['test.auth.refresh_time'] = 3000
-        assert create_context_variable(grizzly, 'test.value', '$conf::test.auth.refresh_time') == {
-            'test': {
-                'value': 3000,
-            }
-        }
-    finally:
-        GrizzlyContext.destroy()
-        del os.environ['HELLO_WORLD']
-
-
-def test_resolve_variable() -> None:
-    grizzly = GrizzlyContext()
-
-    try:
-        assert 'test' not in grizzly.state.variables
-        with pytest.raises(AssertionError):
-            resolve_variable(grizzly, '{{ test }}')
-
-        grizzly.state.variables['test'] = 'some value'
-        assert resolve_variable(grizzly, '{{ test }}') == 'some value'
-
-        assert resolve_variable(grizzly, "now | format='%Y-%m-%d %H'") == "now | format='%Y-%m-%d %H'"
-
-        assert resolve_variable(grizzly, "{{ test }} | format='%Y-%m-%d %H'") == "some value | format='%Y-%m-%d %H'"
-
-        assert resolve_variable(grizzly, 'static value') == 'static value'
-        assert resolve_variable(grizzly, '"static value"') == 'static value'
-        assert resolve_variable(grizzly, "'static value'") == 'static value'
-        assert resolve_variable(grizzly, "'static' value") == "'static' value"
-        assert resolve_variable(grizzly, "static 'value'") == "static 'value'"
-
-        with pytest.raises(ValueError):
-            resolve_variable(grizzly, "'static value\"")
-
-        with pytest.raises(ValueError):
-            resolve_variable(grizzly, "static 'value\"")
-
-        with pytest.raises(ValueError):
-            resolve_variable(grizzly, "'static\" value")
-
-        grizzly.state.variables['number'] = 100
-        assert resolve_variable(grizzly, '{{ (number * 0.25) | int }}') == 25
-
-        assert resolve_variable(grizzly, '{{ (number * 0.25 * 0.2) | int }}') == 5
-
-        try:
-            with pytest.raises(AssertionError):
-                resolve_variable(grizzly, '$env::HELLO_WORLD')
-
-            os.environ['HELLO_WORLD'] = 'first environment variable!'
-
-            assert resolve_variable(grizzly, '$env::HELLO_WORLD') == 'first environment variable!'
-
-            os.environ['HELLO_WORLD'] = 'first "environment" variable!'
-            assert resolve_variable(grizzly, '$env::HELLO_WORLD') == 'first "environment" variable!'
-        finally:
-            del os.environ['HELLO_WORLD']
-
-        with pytest.raises(AssertionError):
-            resolve_variable(grizzly, '$conf::sut.host')
-
-        grizzly.state.configuration['sut.host'] = 'http://host.docker.internal:8003'
-
-        assert resolve_variable(grizzly, '$conf::sut.host')
-
-        grizzly.state.configuration['sut.greeting'] = 'hello "{{ test }}"!'
-        assert resolve_variable(grizzly, '$conf::sut.greeting') == 'hello "{{ test }}"!'
-
-        with pytest.raises(ValueError):
-            resolve_variable(grizzly, '$test::hello')
-
-        assert resolve_variable(grizzly, '') == ''
-    finally:
-        GrizzlyContext.destroy()
-
-
-def test__objectify() -> None:
-    testdata: Dict[str, Any] = {
-        'AtomicInteger': {
-            'test': 1337,
-        },
-        'test': 1338,
-        'AtomicCsvRow': {
-            'input': {
-                'test1': 'hello',
-                'test2': 'world!',
-            }
-        },
-        'Test': {
-            'test1': {
-                'test2': {
-                    'test3': 'value',
-                }
-            }
-        }
-    }
-
-    obj = _objectify(testdata)
-
-    assert (
-        obj['AtomicInteger'].__module__ == 'grizzly.utils' and
-        obj['AtomicInteger'].__class__.__name__ == 'Testdata'
-    )
-    assert getattr(obj['AtomicInteger'], 'test') == 1337
-    assert isinstance(obj['test'], int)
-    assert obj['test'] == 1338
-    assert (
-        obj['AtomicCsvRow'].__module__ == 'grizzly.utils' and
-        obj['AtomicCsvRow'].__class__.__name__ == 'Testdata'
-    )
-    atomiccsvrow_input = getattr(obj['AtomicCsvRow'], 'input', None)
-    assert atomiccsvrow_input is not None
-    assert (
-        atomiccsvrow_input.__module__ == 'grizzly.utils' and
-        atomiccsvrow_input.__class__.__name__ == 'Testdata'
-    )
-    assert getattr(atomiccsvrow_input, 'test1', None) == 'hello'
-    assert getattr(atomiccsvrow_input, 'test2', None) == 'world!'
-
-    assert (
-        obj['Test'].__module__ == 'grizzly.utils' and
-        obj['Test'].__class__.__name__ == 'Testdata'
-    )
-    test = getattr(obj['Test'], 'test1', None)
-    assert test is not None
-    assert (
-        test.__module__ == 'grizzly.utils' and
-        test.__class__.__name__ == 'Testdata'
-    )
-    test = getattr(test, 'test2', None)
-    assert test is not None
-    assert (
-        test.__module__ == 'grizzly.utils' and
-        test.__class__.__name__ == 'Testdata'
-    )
-    test = getattr(test, 'test3', None)
-    assert test is not None
-    assert isinstance(test, str)
-    assert test == 'value'
-
-
-def test_transform_raw() -> None:
-    data = {
-        'test.number.value': 1337,
-        'test.number.description': 'simple description',
-        'test.string.value': 'hello world!',
-        'test.bool.value': True,
-    }
-
-    actual = transform(data, True)
-
-    assert actual == {
-        'test': {
-            'number': {
-                'value': 1337,
-                'description': 'simple description',
-            },
-            'string': {
-                'value': 'hello world!',
-            },
-            'bool': {
-                'value': True,
-            }
-        }
-    }
-
-
-@pytest.mark.usefixtures('cleanup')
-def test_transform(cleanup: Callable) -> None:
-    try:
-        data: Dict[str, Any] = {
-            'AtomicInteger.test': 1337,
-            'test': 1338,
-            'AtomicCsvRow.input.test1': 'hello',
-            'AtomicCsvRow.input.test2': 'world!',
-            'Test.test1.test2.test3': 'value',
-            'Test.test1.test2.test4': 'value',
-            'Test.test2.test3': 'value',
-        }
-
-        obj = transform(data)
-
-        assert (
-            obj['AtomicInteger'].__module__ == 'grizzly.utils' and
-            obj['AtomicInteger'].__class__.__name__ == 'Testdata'
-        )
-        assert getattr(obj['AtomicInteger'], 'test', None) == 1337
-        assert isinstance(obj['test'], int)
-        assert obj['test'] == 1338
-        assert (
-            obj['AtomicCsvRow'].__module__ == 'grizzly.utils' and
-            obj['AtomicCsvRow'].__class__.__name__ == 'Testdata'
-        )
-        assert (
-            obj['AtomicCsvRow'].input.__module__ == 'grizzly.utils' and
-            obj['AtomicCsvRow'].input.__class__.__name__ == 'Testdata'
-        )
-        assert getattr(obj['AtomicCsvRow'].input, 'test1', None) == 'hello'
-        assert getattr(obj['AtomicCsvRow'].input, 'test2', None) == 'world!'
-        assert (
-            obj['Test'].__module__ == 'grizzly.utils' and
-            obj['Test'].__class__.__name__ == 'Testdata'
-        )
-        test = getattr(obj['Test'], 'test1', None)
-        assert test is not None
-        assert (
-            test.__module__ == 'grizzly.utils' and
-            test.__class__.__name__ == 'Testdata'
-        )
-        test = getattr(test, 'test2', None)
-        assert test is not None
-        assert (
-            test.__module__ == 'grizzly.utils' and
-            test.__class__.__name__ == 'Testdata'
-        )
-        test = getattr(test, 'test3', None)
-        assert test is not None
-        assert isinstance(test, str)
-        assert test == 'value'
-    finally:
-        cleanup()
 
 
 def test_in_correct_section() -> None:
