+++ tests/test_grizzly/testdata/variables/test_csv_row.py
@@ -7,7 +7,8 @@ import pytest
 
 from _pytest.tmpdir import TempdirFactory
 
-from grizzly.testdata.variables import AtomicCsvRow, atomiccsvrow__base_type__
+from grizzly.testdata.variables import AtomicCsvRow
+from grizzly.testdata.variables.csv_row import atomiccsvrow__base_type__
 
 from ..fixtures import cleanup  # pylint: disable=unused-import
